# docs/swagger.yaml
definitions:
  Dto.Avatar:
    properties:
      color:
        description: hex color code
        type: string
      hash:
        type: string
      kind:
        description: e.g., "user", "org"
        type: string
      name:
        type: string
    type: object
...
  Dto.CreateRepository:
    properties:
      description:
        type: string
      is_public:
        type: boolean
      kind:
        type: string
      name:
        type: string
      namespace:
        description: Can be null if is a global repo, not organization scoped
        type: string
    type: object
...
  Dto.Repository:
    properties:
      description:
        type: string
      is_public:
        type: boolean
      is_starred:
        type: boolean
      kind:
        description: Image or Application
        type: string
      name:
        description: Image name
        type: string
      namespace:
        type: string
      state:
        type: string
    type: object
  Dto.RepositoryDetails:
    properties:
      can_admin:
        type: boolean
      can_write:
        type: boolean
      description:
        type: string
      is_organization:
        type: boolean
      is_public:
        type: boolean
      is_starred:
        type: boolean
      kind:
        type: string
      name:
        type: string
      namespace:
        type: string
      state:
        type: string
      stats:
        items:
          $ref: '#/definitions/Dto.RepositoryStats'
        type: array
      status_token:
        type: string
      tag_expiration_s:
        type: integer
      tags:
        items:
          $ref: '#/definitions/Dto.RepositoryTag'
        type: array
      trust_enabled:
        type: boolean
    type: object
  Dto.RepositoryPermission:
    properties:
      avatar:
        $ref: '#/definitions/Dto.Avatar'
      isRobot:
        description: Optional fields
        type: boolean
      name:
        type: string
      role:
        type: string
    type: object
  Dto.RepositoryStats:
    properties:
      count:
        type: integer
      date:
        description: 'In format: yyyy-mm-dd'
        type: string
    type: object
  Dto.RepositoryTag:
    properties:
      last_modified:
        type: string
      manifest_digest:
        type: string
      name:
        type: string
      size:
        description: In bits
        type: integer
    type: object
...
  Dto.UpdateRepository:
    properties:
      description:
        type: string
    type: object
  Dto.UpdateRepositoryPermission:
    properties:
      role:
        description: 'allowed value: "admin", "write", "read"'
        type: string
    type: object
...
  Errors.ErrorDetails:
    properties:
      code:
        description: Code is a sting that represents the error code (e.g. "USER_NOT_FOUND",
          "INVALID_INPUT", etc.)
        type: string
      message:
        type: string
    type: object
  Errors.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/Errors.ErrorDetails'
    type: object
info:
  contact: {}
  description: Quay registry API implemented in Go
  title: Quay Go API
  version: "1.0"
paths:
...
  /api/v1/repository:
    get:
      description: List repositories
      parameters:
      - description: Organization or Username as namespace to filter repositories
        in: query
        name: namespace
        type: string
      - description: Filter on public repositories
        in: query
        name: is_public
        type: boolean
      - description: Filter on stared repositories
        in: query
        name: is_starred
        type: boolean
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Dto.Repository'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List repositories
      tags:
      - Repository
    post:
      consumes:
      - application/json
      description: Create a repository
      parameters:
      - description: Repository metadata
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/Dto.CreateRepository'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Dto.Repository'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a repository
      tags:
      - Repository
  /api/v1/repository/{repository}:
    delete:
      consumes:
      - application/json
      description: Delete a repository
      parameters:
      - description: Name of the repository
        in: path
        name: repository
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a repository
      tags:
      - Repository
    get:
      description: Get a repository
      parameters:
      - description: Repository name in the format namespace/repository
        in: path
        name: repository
        required: true
        type: string
      - description: Include tags in the repository response
        in: query
        name: include_tags
        type: boolean
      - description: Include statistics in the repository response
        in: query
        name: include_stats
        type: boolean
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Dto.RepositoryDetails'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a repository
      tags:
      - Repository
    patch:
      consumes:
      - application/json
      description: Update repository details
      parameters:
      - description: Name of the repository
        in: path
        name: repository
        required: true
        type: string
      - description: Repository details to change
        in: body
        name: update
        required: true
        schema:
          $ref: '#/definitions/Dto.UpdateRepository'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Dto.Repository'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update repository details
      tags:
      - Repository
  /api/v1/repository/{repository}/permissions/team:
    get:
      description: List teams permission on a repository
      parameters:
      - description: Repository name in the format namespace/repository
        in: path
        name: repository
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Dto.RepositoryPermission'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List teams permission on a repository
      tags:
      - Permission
  /api/v1/repository/{repository}/permissions/team/{teamname}:
    get:
      description: Get a team's permission on a repository
      parameters:
      - description: Repository name in the format namespace/repository
        in: path
        name: repository
        required: true
        type: string
      - description: Team name
        in: path
        name: teamname
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Dto.RepositoryPermission'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a team's permission on a repository
      tags:
      - Permission
    patch:
      consumes:
      - application/json
      description: Update team repository permission
      parameters:
      - description: Repository name in the format namespace/repository
        in: path
        name: repository
        required: true
        type: string
      - description: Team name
        in: path
        name: teamname
        required: true
        type: string
      - description: Repository details to change
        in: body
        name: update
        required: true
        schema:
          $ref: '#/definitions/Dto.UpdateRepositoryPermission'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Dto.Repository'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Errors.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update team repository permission
      tags:
      - Permission
...
securityDefinitions:
  ApiKeyAuth:
    description: API key authentication using the Authorization header. The value
      should be in the format "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"